Made the old file handling tasks even more deprecated. Made Ant use the new tasks in its own buildfile.
Fixed a problem with <delete dir=""> when the directory didn't exists
(deltree would ignore this - delete now does so as well).


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@268052 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/Delete.java
@@ -228,7 +228,7 @@ public class Delete extends MatchingTask {
         }
 
         // delete the directory
-        if (dir != null && !usedMatchingTask) {
+        if (dir != null && dir.exists() && dir.isDirectory() && !usedMatchingTask) {
             log("Deleting directory " + dir.getAbsolutePath());
             removeDir(dir);
         }
